# Optimize creating bytes out of Mappair