### What's being changed:

Two commits:
1. optimizes the .Bytes function of MapPair (stats are for this optimization) reducing runtime and allocs
  a. The optimization should reduce allocations for every use of `NewReadWriter`
2. Moves some code before the lock to reduce the amount of time the lock needs to be held (no stats, but I don't think this is bad in any case)

Before
```
BenchmarkBytes/10	         7966041	       138.5 ns/op	      64 B/op	       1 allocs/op
BenchmarkBytes/100        	 6687142	       176.6 ns/op	     192 B/op	       2 allocs/op
BenchmarkBytes/1000       	 4867688	       243.4 ns/op	    1088 B/op	       2 allocs/op
BenchmarkBytes/10000      	 1432347	       815.8 ns/op	   10304 B/op	       2 allocs/op
BenchmarkBytes/24576      	  725058	      1431 ns/op	   27328 B/op	       2 allocs/op
```

After
```
BenchmarkBytes/10	         8877076	       129.9 ns/op	      24 B/op	       1 allocs/op
BenchmarkBytes/100        	 8969838	       129.4 ns/op	     112 B/op	       1 allocs/op
BenchmarkBytes/1000       	 5925535	       198.7 ns/op	    1024 B/op	       1 allocs/op
BenchmarkBytes/10000      	 1665730	       714.3 ns/op	   10240 B/op	       1 allocs/op
BenchmarkBytes/24576      	  788707	      1312 ns/op	   27264 B/op	       1 allocs/op
```

### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [ ] All new code is covered by tests where it is reasonable.
- [ ] Performance tests have been run or not necessary.

<!-- Uncomment the following section if this PR requires changes in related projects (e.g., documentation, client libraries).

GitHub actions will automatically create an issue in the corresponding repository for each checked box below. (See `.github/workflows/create-cross-functional-issues.yml`)

### Cross-functional impact

- [ ] This change requires public documentation (weaviate-io) to be updated. Check the box to automatically create a corresponding issue.
- Does it require a change in the client libraries? If yes, please check the boxes for the affected client libraries.
    - [ ] Python (weaviate-python-client)
    - [ ] JavaScript/TypeScript (typescript-client)
    - [ ] Go (weaviate-go-client)
    - [ ] Java (java-client)

-->